refactor(enrollment): extraire le filtre statut du SearchForm
